refactor: move common methods from SNS_source_cpu to _SNS_source_utils
