+++ tests/components/sources/test_SNS_source_cpu.py
@@ -5,7 +5,8 @@ thisdir = os.path.dirname(__file__)
 dat = os.path.join(thisdir, 'source_rot2_cdr_cyl_3x3_20190417.dat')
 
 from mcni import neutron_buffer, neutron
-from mcvine.acc.components.sources.SNS_source_cpu import SNS_source, init, generate
+from mcvine.acc.components.sources.SNS_source_cpu import SNS_source, generate
+from mcvine.acc.components.sources._SNS_source_utils import init
 
 def test_generate():
     INorm2, Es, Pvec, ts, Ptmat, EPmin, EPmax, Eidx_range, tidx_range = init(3, 1500., dat)
@@ -34,7 +35,7 @@ def test_component_n1e6():
     return
 
 def main():
-    # test_generate()
+    test_generate()
     # test_component()
     test_component_n1e6()
     return
